moved new unit tests (former regression tests) to correct location in joshua 7 (joshua-core/src)
