+++ joshua-core/src/test/java/org/apache/joshua/decoder/cky/BnEnDecodingTest.java
@@ -116,6 +116,6 @@ public class BnEnDecodingTest {
 	public void configureDecoder(String pathToConfig) throws Exception {
 		joshuaConfig = new JoshuaConfiguration();
 		joshuaConfig.readConfigFile(pathToConfig);
-		KenLmTestUtil.Guard(() -> decoder = new Decoder(joshuaConfig, ""));
+		KenLmTestUtil.Guard(() -> decoder = new Decoder(joshuaConfig));
 	}
 }
