+++ joshua-core/src/test/java/org/apache/joshua/decoder/cky/LeftStateTest.java
@@ -60,6 +60,6 @@ public class LeftStateTest {
 	public void configureDecoder(String pathToConfig) throws Exception {
 		joshuaConfig = new JoshuaConfiguration();
 		joshuaConfig.readConfigFile(pathToConfig);
-		KenLmTestUtil.Guard(() -> decoder = new Decoder(joshuaConfig, ""));
+		KenLmTestUtil.Guard(() -> decoder = new Decoder(joshuaConfig));
 	}
 }
