+++ joshua-core/src/test/java/org/apache/joshua/decoder/cky/NAryTest.java
@@ -59,6 +59,6 @@ public class NAryTest {
   public void configureDecoder(String pathToConfig) throws Exception {
     joshuaConfig = new JoshuaConfiguration();
     joshuaConfig.readConfigFile(pathToConfig);
-    KenLmTestUtil.Guard(() -> decoder = new Decoder(joshuaConfig, ""));
+    KenLmTestUtil.Guard(() -> decoder = new Decoder(joshuaConfig));
   }
 }
