+++ joshua-core/src/test/resources/decoder/lowercaser/joshua.config
