+++ joshua-core/src/test/resources/phrase_decoder/rules.packed/encoding
