+++ joshua-core/src/test/resources/phrase_decoder/rules.packed/slice_00000.target
